+++ release/releasenotes.md
@@ -9,6 +9,7 @@
  - stupid typo in IsInRange()
  - decoupled old/global gameinfo
  - small texts update
+ - consequent use of wxString::IsEmpty() i.s.o wxString::empty()
 
 #V10.4.0 Tuesday 4 March 2025
  - updated AutoHotkey64.exe from version 2.0.18 to 2.0.19